requirements.txt
lxml
mwxml
mysql-connector-python
pymenu-console
openpyxl
